+++ CHANGELOG.md
@@ -2,6 +2,14 @@ Change Log
 ==========
 
 
+Version 1.2.0 *(2014-10-21)*
+----------------------------
+
+* Support multiple iconic fonts.
+* Support for displaying the preview in edit mode.
+* Fix: Font size too large to fit in cache.
+
+
 Version 1.1.0 *(2014-10-06)*
 ----------------------------